Switched to size_T for all indexs in Testing

+++ Race/Testing.cpp
@@ -12,7 +12,7 @@ double NeunzigGradSteigung = 5728; //[%] 89° da 90° undefiniert
 string IDStartpunkt = "6535972824";
 double LatitudeStartPunkt = 50.3451467;
 double LongitudeStartPunkt = 6.9264464;
-double MaxSpeedTesla = 261 /3.6;  //[m/s]
+double MaxSpeedTesla = 261 / 3.6;  //[m/s]
 
 void ErrorLog::TestDatenbeschaffung(vector<node>& nodes) {
 	Testing Test2add;
@@ -23,7 +23,7 @@ void ErrorLog::TestDatenbeschaffung(vector<node>& nodes) {
 		Test2add.Ergebnisse = "Failed";
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	Test2add.Aufgabe = "Datenbeschaffung";
 	Test2add.Testname = "KoordinatenStartPunkt";
 	Test2add.Ergebnisse = "Passed";
@@ -34,7 +34,7 @@ void ErrorLog::TestDatenbeschaffung(vector<node>& nodes) {
 }
 
 void ErrorLog::TestDatenAufbereitung(vector<node>& nodes) {
-	int index = 0;
+	size_t index = 0;
 	Testing Test2add;
 	Test2add.Aufgabe = "Datenaufbereitung";
 	Test2add.Testname = "SteigungKleiner90Grad";
@@ -47,7 +47,7 @@ void ErrorLog::TestDatenAufbereitung(vector<node>& nodes) {
 		index++;
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	Test2add.Aufgabe = "Datenaufbereitung";
 	Test2add.Testname = "SteigungWertebereich";
 	Test2add.Ergebnisse = "Passed";
@@ -59,7 +59,7 @@ void ErrorLog::TestDatenAufbereitung(vector<node>& nodes) {
 		index++;
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	Test2add.Aufgabe = "Datenaufbereitung";
 	Test2add.Testname = "MinKurvenRadius";
 	Test2add.Ergebnisse = "Passed";
@@ -71,7 +71,7 @@ void ErrorLog::TestDatenAufbereitung(vector<node>& nodes) {
 		index++;
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	Test2add.Aufgabe = "Datenaufbereitung";
 	Test2add.Testname = "DistanceToNext";
 	Test2add.Ergebnisse = "Passed";
@@ -86,7 +86,7 @@ void ErrorLog::TestDatenAufbereitung(vector<node>& nodes) {
 }
 
 void ErrorLog::TestSollfahrtbestimmung(vector<node>& nodes) {
-	int index = 0;
+	size_t index = 0;
 	Testing Test2add;
 	Test2add.Aufgabe = "Sollfahrtbestimmung";
 	Test2add.Testname = "Vmax kleiner Vmax_Limited";
@@ -99,7 +99,7 @@ void ErrorLog::TestSollfahrtbestimmung(vector<node>& nodes) {
 		index++;
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	index = 0;
 	Test2add.Aufgabe = "Sollfahrtbestimmung";
 	Test2add.Testname = "Nur positive Geschwindigkeit";
@@ -122,7 +122,7 @@ void ErrorLog::TestFahrphysik(vector<node>& nodes) {
 		Test2add.Ergebnisse = "Failed";
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	Test2add.Aufgabe = "Fahrphysik";
 	Test2add.Testname = "KoordinatenStartPunkt";
 	Test2add.Ergebnisse = "Passed";
@@ -130,7 +130,7 @@ void ErrorLog::TestFahrphysik(vector<node>& nodes) {
 		Test2add.Ergebnisse = "Failed";
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	Test2add.Aufgabe = "Fahrphysik";
 	Test2add.Testname = "GeschwindigkeitStartPunkt";
 	Test2add.Ergebnisse = "Passed";
@@ -138,8 +138,8 @@ void ErrorLog::TestFahrphysik(vector<node>& nodes) {
 		Test2add.Ergebnisse = "Failed";
 	}
 	Testvektor.push_back(Test2add);
-	
-	int index = 0;
+
+	size_t index = 0;
 	Test2add.Aufgabe = "Fahrphysik";
 	Test2add.Testname = "KeinNaNmehr";
 	Test2add.Ergebnisse = "Passed";
@@ -160,7 +160,7 @@ void ErrorLog::TestFahrphysik(vector<node>& nodes) {
 		index++;
 	}
 	Testvektor.push_back(Test2add);
-	
+
 	index = 0;
 	Test2add.Aufgabe = "Fahrphysik";
 	Test2add.Testname = "MaxSpeedRealistisch";
